requirements.txt
dash 
yfinance
pandas
numpy
google_generativeai == 0.8.4